dxe_core: Add MemoryProtectionPolicy struct
This commit centralizes all memory protections policy
decisions under a single struct impl for easy auditability
and ensuring of consistency between different environments.

This structure is well documented and unit tested. All current
policy decision makers are updated to use this structure instead.

The GCD owns the single instance of this struct because it relies
on the default protection attributes, which may be changed if
compatibility mode is entered. All other usage of the struct is
through associated functions. They do not rely on any state.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>