patina_debugger: Switch breakpoint to be conditional (#1087)
The breakpoint always causing an exception may not be well understood
and could lead to people leaving breakpoints in production code causing
unexpected exceptions. This commit changes the existing `breakpoint`
function to only trigger a breakpoint if the debugger is enabled, and
renames the existing `breakpoint` function to `breakpoint_unchecked` to
indicate that it will always trigger a breakpoint regardless of the
debugger state.