patina_dxe_core: Prevent setting hob_list from being optimized out (#1106)
## Description

The compiler, on release builds, is optimizing out the `call_once` call
in `set_hob_list`. This resulted in future calls to `hob_list` to panic.
This commit changes the debug_assert to an assert and updated
`set_hob_list` to use the output returned from `call_once` so that it is
not optimized away.

REF: https://github.com/OpenDevicePartnership/patina-qemu/pull/86

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Boot to shell Q35 RELEASE

## Integration Instructions

N/A