patina_dxe_core: Clean up HOB init flow (#1120)
## Description

The HOB list is relocated to DXE allocated memory shortly into DXE core
entry. The Rust allocated HOB list is not binary compatible with the C
HOB list. The Rust HOB list is a vector of references to HOB structures
while the C hob list is a compact series of binary HOB structures.

While the Rust HOB list is sufficient for most use cases that simply
need information from the HOBs in a more type safe and ergonomic
interface, there are still use cases that require binary compatibility
of the HOB list such as placement of the HOB into the system table where
it is discovered by other modules, including C modules.

Today, this works because the HOB list pointer is directly passed to the
system table initialization function where it is placed into the HOB
list system table. This change relocates both HOB lists in the same
place in the `init_memory()` function. Later DXE functionality
consistently works on a relocated HOB list after `init_memory()`
returns.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- `cargo make all`
- QEMU SBSA and Q35 EFI shell boot

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>