patina_dxe_core: PI dispatcher subsystem / removal of statics (#1121)
## Description

This pull-request combines `fv.rs` and `dispatcher.rs` into a single
module (`pi_dispatcher`) and removes the statics currently used in
`fv.rs` and `dispatcher.rs`. 

It additionally re-works `fv.rs` to have stronger boundaries between the pure rust protocol functions and the EFIAPI wrapper methods. This includes catching a bug that would result in a panic (Buffer overflow) when calling `fv_read_section` when the provided buffer size was greater than the section size. A test was written for this scenario to prevent regression.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI continues to pass. Boot to shell in Q35.

## Integration Instructions

N/A