[AArch64] Fixing a SPI interrupt ID miscalculation (#1156)
## Description

The SPI interrupt ID range comes after the PPI range, which starts from
32. The current logic compensates this by subtracting the the starting
index, however the subtraction expression is incorrect.

This change fixed the index miscalculation.

In addition, it also removes 2 redundant steps for setting the priority
as it is already covered in gic setup step.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

This was tested on both QEMU SBSA and proprietary hardware platform,
verified that the ARM watchdog timer can fire properly through the
corresponding interrupt handler.

## Integration Instructions

N/A