patina_dxe_core: image.rs: Improve usage of safe abstractions (#1177)
## Description

This pull request updates `image.rs` to use more rust abstractions over
the more low-level / raw pointer usage we were using before. While this
pull request will be squashed, each commit is cleanly broken out into
it's own change. **It is suggested to review this PR commit by commit**.

**Commit 1:** This commit updates the hii_resource_section to be stored
in an `Option<Box<[u8]>>` created via the `CoreMemoryManager` and the
`MemoryManager` trait rather than a `Option<*mut [u8]>`. This allows us
to remove multiple fields from `PrivateImageData` that existed purely to
manually free the pages when dropped and simplify the management of this
field.

**Commit 2:**: This commit updates the loaded image buffer to be stored
in a `Box<[u8]>` created via the `CoreMemoryManager` and the
`MemoryManager` trait for the same reasons as Commit 1.

**Commit 3:** This commit updates the `ImageStack` struct to be a
`Box<[u8]>` created via the `CoreMemoryManager` and the `MemoryManager`
trait for the same reasons as Commit 1.

**Commit 4:** This commit removes the unnecessary `image_info_ptr`
field, which is just a pointer to the ` image_info:
Box<efi::protocols::loaded_image::Protocol>,` field. We instead get the
pointer directly from the Box.

**Commit 5:** This commit updates the `image_device_path_ptr` to store
an `Option<Box<[u8]>>` rather than a `*mut c_void` and cleanly manages
the lifetime of this data rather then leaking it. **NOTE: This commit
has a bugfix to re-insert the private_image_data if we fail to uninstall
either the `loaded_image` or `loaded_image_device_path` protocols for a
reason other than NotFound**.

**Commit 6**: This commit simply pulls out the `exit_data` as a new
struct so that we can mark `Send + Sync` on only that field. This is to
eventually be able to remove the blanket `Send + Sync` implementation on
the `DxeCoreGlobalImageData` struct.

**Commit 7:** Adds tests

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

All CI passes. Continue to boot to shell on Q35

## Integration Instructions

N/A