image: do not finish loading image if platform policy prevents it. (#1191)
## Description

In the EDKII implementation of the DXE Core, load_image will only
continue loading an image when authentication fails if the error is a
EFI_SECURITY_VIOLATION error. Any other error (and specifically
EFI_ACCESS_DENIED) will immediately abort without actually loading the
image.

The Patina DXE Core loads the image regardless of the result of the
authentication, and only reports the result to the caller after the load
has completed. This appears to be a security risk because even though
the handle is returned as null, the protocols are still loaded and
accessible. A call to `LocateHandleBuffer` could get you the handle and
then a `StartImage` could start it.

This commit stops the loading of an image if authentication returns any
error other then EFI_SECURITY_VIOLATION.

- [x] Impacts functionality?
- [x] Impacts security?
- [ ] Breaking change?
- [x] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

CI, Boot to shell.

## Integration Instructions

N/A